// src/api/generated/iMSIncidentManagementSystem.schemas.ts
export interface LocationCreate {
  name: string;
  location_type_id: string;
  address?: string | null;
}

export interface LocationTypeCreate {
  name: string;
}

export interface LocationTypeUpdate {
  name?: string | null;
}

export interface LocationUpdate {
  name?: string | null;
  location_type_id?: string | null;
  address?: string | null;
}